FW, math and modular

+++ math/modular.cpp
@@ -1,13 +1,13 @@
-const int mod = 1e9 + 7;
-template <int MOD=mod> struct Modular {
-	int value;
-	static const int MOD_value = MOD;
-	Modular(long long v = 0) { value = v % MOD; if (value < 0) value += MOD;}
-	Modular(long long a, long long b) : value(0){ *this += a; *this /= b;}
+const lli mod = 1e9 + 7;
+template <lli MOD=mod> struct Modular {
+	lli value;
+	static const lli MOD_value = MOD;
+	Modular(lli v = 0) { value = v % MOD; if (value < 0) value += MOD;}
+	Modular(lli a, lli b) : value(0){ *this += a; *this /= b;}
 	Modular& operator+=(Modular const& b) {value += b.value; if (value >= MOD) value -= MOD; return *this;}
 	Modular& operator-=(Modular const& b) {value -= b.value; if (value < 0) value += MOD;return *this;}
-	Modular& operator*=(Modular const& b) {value = (long long)value * b.value % MOD;return *this;}
-	friend Modular powmod(Modular a, long long e) {
+	Modular& operator*=(Modular const& b) {value = (lli)value * b.value % MOD;return *this;}
+	friend Modular powmod(Modular a, lli e) {
 		Modular res = 1; while (e) { if (e&1) res *= a; a *= a; e >>= 1; }
 		return res;
 	}
